server: make conn.serveImpl a Server method
conn.serveImpl is only called from one place and uses 3 fields from Server so
making it a method on Server reduces number of params.

Release note: None
