+++ pkg/sql/pgwire/server.go
@@ -15,6 +15,7 @@ import (
 	"context"
 	"io"
 	"net"
+	"sync/atomic"
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/base"
@@ -23,6 +24,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/settings"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql"
+	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/hba"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/identmap"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
@@ -37,6 +39,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/log/logpb"
 	"github.com/cockroachdb/cockroach/pkg/util/metric"
 	"github.com/cockroachdb/cockroach/pkg/util/mon"
+	"github.com/cockroachdb/cockroach/pkg/util/netutil"
 	"github.com/cockroachdb/cockroach/pkg/util/stop"
 	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
@@ -830,15 +833,10 @@ func (s *Server) ServeConn(
 		connStart,
 	)
 
-	var afterReadMsgTestingKnob func(context.Context) error
-	if s.execCfg.PGWireTestingKnobs != nil {
-		afterReadMsgTestingKnob = s.execCfg.PGWireTestingKnobs.AfterReadMsgTestingKnob
-	}
 	// Do the reading of commands from the network.
-	c.serveImpl(
+	s.serveImpl(
 		ctx,
-		s.IsDraining,
-		s.SQLServer,
+		c,
 		&tenantReserved,
 		authOptions{
 			connType:        preServeStatus.ConnType,
@@ -848,7 +846,6 @@ func (s *Server) ServeConn(
 			identMap:        identMap,
 			testingAuthHook: testingAuthHook,
 		},
-		afterReadMsgTestingKnob,
 	)
 	return nil
 }
@@ -873,6 +870,410 @@ func (s *Server) newConn(netConn net.Conn, sArgs sql.SessionArgs, connStart time
 	c.errWriter.sv = sv
 	c.errWriter.msgBuilder = &c.msgBuilder
 	return c
+}
+
+// maxRepeatedErrorCount is the number of times an error can be received
+// while reading from the network connection before the server decides to give
+// up and abort the connection.
+const maxRepeatedErrorCount = 1 << 15
+
+// serveImpl continuously reads from the network connection and pushes execution
+// instructions into a sql.StmtBuf, from where they'll be processed by a command
+// "processor" goroutine (a connExecutor).
+// The method returns when the pgwire termination message is received, when
+// network communication fails, when the server is draining or when ctx is
+// canceled (which also happens when draining (but not from the get-go), and
+// when the processor encounters a fatal error).
+//
+// serveImpl always closes the network connection before returning.
+//
+// sqlServer is used to create the command processor. As a special facility for
+// tests, sqlServer can be nil, in which case the command processor and the
+// write-side of the connection will not be created.
+//
+// Internally, a connExecutor will be created to execute commands. Commands read
+// from the network are buffered in a stmtBuf which is consumed by the
+// connExecutor. The connExecutor produces results which are buffered and
+// sometimes synchronously flushed to the network.
+//
+// The reader goroutine (this one) outlives the connExecutor's goroutine (the
+// "processor goroutine").
+// However, they can both signal each other to stop. Here's how the different
+// cases work:
+// 1) The reader receives a ClientMsgTerminate protocol packet: the reader
+// closes the stmtBuf and also cancels the command processing context. These
+// actions will prompt the command processor to finish.
+// 2) The reader gets a read error from the network connection: like above, the
+// reader closes the command processor.
+// 3) The reader's context is canceled (happens when the server is draining but
+// the connection was busy and hasn't quit yet): the reader notices the canceled
+// context and, like above, closes the processor.
+// 4) The processor encounters an error. This error can come from various fatal
+// conditions encountered internally by the processor, or from a network
+// communication error encountered while flushing results to the network.
+// The processor will cancel the reader's context and terminate.
+// Note that query processing errors are different; they don't cause the
+// termination of the connection.
+//
+// Draining notes:
+//
+// The reader notices that the server is draining by polling the IsDraining
+// closure passed to serveImpl. At that point, the reader delegates the
+// responsibility of closing the connection to the statement processor: it will
+// push a DrainRequest to the stmtBuf which signals the processor to quit ASAP.
+// The processor will quit immediately upon seeing that command if it's not
+// currently in a transaction. If it is in a transaction, it will wait until the
+// first time a Sync command is processed outside of a transaction - the logic
+// being that we want to stop when we're both outside transactions and outside
+// batches.
+func (s *Server) serveImpl(
+	ctx context.Context, c *conn, reserved *mon.BoundAccount, authOpt authOptions,
+) {
+	defer func() { _ = c.conn.Close() }()
+
+	if c.sessionArgs.User.IsRootUser() || c.sessionArgs.User.IsNodeUser() {
+		ctx = logtags.AddTag(ctx, "user", redact.Safe(c.sessionArgs.User))
+	} else {
+		ctx = logtags.AddTag(ctx, "user", c.sessionArgs.User)
+	}
+	tracing.SpanFromContext(ctx).SetTag("user", attribute.StringValue(c.sessionArgs.User.Normalized()))
+
+	sqlServer := s.SQLServer
+	inTestWithoutSQL := sqlServer == nil
+	if !inTestWithoutSQL {
+		sessionStart := timeutil.Now()
+		defer func() {
+			if c.authLogEnabled() {
+				endTime := timeutil.Now()
+				ev := &eventpb.ClientSessionEnd{
+					CommonEventDetails:      logpb.CommonEventDetails{Timestamp: endTime.UnixNano()},
+					CommonConnectionDetails: authOpt.connDetails,
+					Duration:                endTime.Sub(sessionStart).Nanoseconds(),
+				}
+				log.StructuredEvent(ctx, ev)
+			}
+		}()
+	}
+
+	// NOTE: We're going to write a few messages to the connection in this method,
+	// for the handshake. After that, all writes are done async, in the
+	// startWriter() goroutine.
+
+	ctx, cancelConn := context.WithCancel(ctx)
+	defer cancelConn() // This calms the linter that wants these callbacks to always be called.
+
+	var sentDrainSignal bool
+	// The net.Conn is switched to a conn that exits if the ctx is canceled.
+	c.conn = NewReadTimeoutConn(c.conn, func() error {
+		// If the context was canceled, it's time to stop reading. Either a
+		// higher-level server or the command processor have canceled us.
+		if ctx.Err() != nil {
+			return ctx.Err()
+		}
+		// If the server is draining, we'll let the processor know by pushing a
+		// DrainRequest. This will make the processor quit whenever it finds a good
+		// time.
+		if !sentDrainSignal && s.IsDraining() {
+			_ /* err */ = c.stmtBuf.Push(ctx, sql.DrainRequest{})
+			sentDrainSignal = true
+		}
+		return nil
+	})
+	c.rd = *bufio.NewReader(c.conn)
+
+	// the authPipe below logs authentication messages iff its auth
+	// logger is non-nil. We define this here.
+	logAuthn := !inTestWithoutSQL && c.authLogEnabled()
+
+	// We'll build an authPipe to communicate with the authentication process.
+	systemIdentity := c.sessionArgs.SystemIdentity
+	if systemIdentity.Undefined() {
+		systemIdentity = c.sessionArgs.User
+	}
+	authPipe := newAuthPipe(c, logAuthn, authOpt, systemIdentity)
+	var authenticator authenticatorIO = authPipe
+
+	// procCh is the channel on which we'll receive the termination signal from
+	// the command processor.
+	var procCh <-chan error
+
+	// We need a value for the unqualified int size here, but it is controlled
+	// by a session variable, and this layer doesn't have access to the session
+	// data. The callback below is called whenever default_int_size changes.
+	// It happens in a different goroutine, so it has to be changed atomically.
+	var atomicUnqualifiedIntSize = new(int32)
+	onDefaultIntSizeChange := func(newSize int32) {
+		atomic.StoreInt32(atomicUnqualifiedIntSize, newSize)
+	}
+
+	if !inTestWithoutSQL {
+		// Spawn the command processing goroutine, which also handles connection
+		// authentication). It will notify us when it's done through procCh, and
+		// we'll also interact with the authentication process through ac.
+		var ac AuthConn = authPipe
+		procCh = c.processCommandsAsync(
+			ctx,
+			authOpt,
+			ac,
+			sqlServer,
+			reserved,
+			cancelConn,
+			onDefaultIntSizeChange,
+		)
+	} else {
+		// sqlServer == nil means we are in a local test. In this case
+		// we only need the minimum to make pgx happy.
+		defer reserved.Close(ctx)
+		var err error
+		for param, value := range testingStatusReportParams {
+			err = c.bufferParamStatus(param, value)
+			if err != nil {
+				break
+			}
+		}
+		if err != nil {
+			return
+		}
+		var ac AuthConn = authPipe
+		// Simulate auth succeeding.
+		ac.AuthOK(ctx)
+		dummyCh := make(chan error)
+		close(dummyCh)
+		procCh = dummyCh
+
+		if err := c.bufferInitialReadyForQuery(0 /* queryCancelKey */); err != nil {
+			return
+		}
+		// We don't have a CmdPos to pass in, since we haven't received any commands
+		// yet, so we just use the initial lastFlushed value.
+		if err := c.Flush(c.writerState.fi.lastFlushed); err != nil {
+			return
+		}
+	}
+
+	var terminateSeen bool
+	var authDone, ignoreUntilSync bool
+	var repeatedErrorCount int
+	for {
+		breakLoop, isSimpleQuery, err := func() (bool, bool, error) {
+			typ, n, err := c.readBuf.ReadTypedMsg(&c.rd)
+			c.metrics.BytesInCount.Inc(int64(n))
+			if err == nil {
+				if knobs := s.execCfg.PGWireTestingKnobs; knobs != nil {
+					if afterReadMsgTestingKnob := knobs.AfterReadMsgTestingKnob; afterReadMsgTestingKnob != nil {
+						err = afterReadMsgTestingKnob(ctx)
+					}
+				}
+			}
+			isSimpleQuery := typ == pgwirebase.ClientMsgSimpleQuery
+			if err != nil {
+				if pgwirebase.IsMessageTooBigError(err) {
+					log.VInfof(ctx, 1, "pgwire: found big error message; attempting to slurp bytes and return error: %s", err)
+
+					// Slurp the remaining bytes.
+					slurpN, slurpErr := c.readBuf.SlurpBytes(&c.rd, pgwirebase.GetMessageTooBigSize(err))
+					c.metrics.BytesInCount.Inc(int64(slurpN))
+					if slurpErr != nil {
+						return false, isSimpleQuery, errors.Wrap(slurpErr, "pgwire: error slurping remaining bytes")
+					}
+				}
+
+				// Write out the error over pgwire.
+				if err := c.stmtBuf.Push(ctx, sql.SendError{Err: err}); err != nil {
+					return false, isSimpleQuery, errors.New("pgwire: error writing too big error message to the client")
+				}
+
+				// If this is a simple query, we have to send the sync message back as
+				// well.
+				if isSimpleQuery {
+					if err := c.stmtBuf.Push(ctx, sql.Sync{
+						// CRDB is implicitly generating this Sync during the simple
+						// protocol.
+						ExplicitFromClient: false,
+					}); err != nil {
+						return false, isSimpleQuery, errors.New("pgwire: error writing sync to the client whilst message is too big")
+					}
+				}
+
+				// We need to continue processing here for pgwire clients to be able to
+				// successfully read the error message off pgwire.
+				//
+				// If break here, we terminate the connection. The client will instead see that
+				// we terminated the connection prematurely (as opposed to seeing a ClientMsgTerminate
+				// packet) and instead return a broken pipe or io.EOF error message.
+				return false, isSimpleQuery, errors.Wrap(err, "pgwire: error reading input")
+			}
+			timeReceived := timeutil.Now()
+			log.VEventf(ctx, 2, "pgwire: processing %s", typ)
+
+			if ignoreUntilSync {
+				if typ != pgwirebase.ClientMsgSync {
+					log.VInfof(ctx, 1, "pgwire: skipping non-sync message after encountering error")
+					return false, isSimpleQuery, nil
+				}
+				ignoreUntilSync = false
+			}
+
+			if !authDone {
+				if typ == pgwirebase.ClientMsgPassword {
+					var pwd []byte
+					if pwd, err = c.readBuf.GetBytes(n - 4); err != nil {
+						return false, isSimpleQuery, err
+					}
+					// Pass the data to the authenticator. This hopefully causes it to finish
+					// authentication in the background and give us an intSizer when we loop
+					// around.
+					if err = authenticator.sendPwdData(pwd); err != nil {
+						return false, isSimpleQuery, err
+					}
+					return false, isSimpleQuery, nil
+				}
+				// Wait for the auth result.
+				if err = authenticator.authResult(); err != nil {
+					// The error has already been sent to the client.
+					return true, isSimpleQuery, nil //nolint:returnerrcheck
+				}
+				authDone = true
+			}
+
+			switch typ {
+			case pgwirebase.ClientMsgPassword:
+				// This messages are only acceptable during the auth phase, handled above.
+				err = pgwirebase.NewProtocolViolationErrorf("unexpected authentication data")
+				return true, isSimpleQuery, c.writeErr(ctx, err, &c.writerState.buf)
+			case pgwirebase.ClientMsgSimpleQuery:
+				if err = c.handleSimpleQuery(
+					ctx, &c.readBuf, timeReceived, parser.NakedIntTypeFromDefaultIntSize(atomic.LoadInt32(atomicUnqualifiedIntSize)),
+				); err != nil {
+					return false, isSimpleQuery, err
+				}
+				return false, isSimpleQuery, c.stmtBuf.Push(ctx, sql.Sync{
+					// CRDB is implicitly generating this Sync during the simple
+					// protocol.
+					ExplicitFromClient: false,
+				})
+
+			case pgwirebase.ClientMsgExecute:
+				// To support the 1PC txn fast path, we peek at the next command to
+				// see if it is a Sync. This is because in the extended protocol, an
+				// implicit transaction cannot commit until the Sync is seen. If there's
+				// an error while peeking (for example, there are no bytes in the
+				// buffer), the error is ignored since it will be handled on the next
+				// loop iteration.
+				followedBySync := false
+				if nextMsgType, err := c.rd.Peek(1); err == nil &&
+					pgwirebase.ClientMessageType(nextMsgType[0]) == pgwirebase.ClientMsgSync {
+					followedBySync = true
+				}
+				return false, isSimpleQuery, c.handleExecute(ctx, &c.readBuf, timeReceived, followedBySync)
+
+			case pgwirebase.ClientMsgParse:
+				return false, isSimpleQuery, c.handleParse(ctx, &c.readBuf, parser.NakedIntTypeFromDefaultIntSize(atomic.LoadInt32(atomicUnqualifiedIntSize)))
+
+			case pgwirebase.ClientMsgDescribe:
+				return false, isSimpleQuery, c.handleDescribe(ctx, &c.readBuf)
+
+			case pgwirebase.ClientMsgBind:
+				return false, isSimpleQuery, c.handleBind(ctx, &c.readBuf)
+
+			case pgwirebase.ClientMsgClose:
+				return false, isSimpleQuery, c.handleClose(ctx, &c.readBuf)
+
+			case pgwirebase.ClientMsgTerminate:
+				terminateSeen = true
+				return true, isSimpleQuery, nil
+
+			case pgwirebase.ClientMsgSync:
+				// We're starting a batch here. If the client continues using the extended
+				// protocol and encounters an error, everything until the next sync
+				// message has to be skipped. See:
+				// https://www.postgresql.org/docs/current/10/protocol-flow.html#PROTOCOL-FLOW-EXT-QUERY
+				return false, isSimpleQuery, c.stmtBuf.Push(ctx, sql.Sync{
+					// The client explicitly sent this Sync as part of the extended
+					// protocol.
+					ExplicitFromClient: true,
+				})
+
+			case pgwirebase.ClientMsgFlush:
+				return false, isSimpleQuery, c.handleFlush(ctx)
+
+			case pgwirebase.ClientMsgCopyData, pgwirebase.ClientMsgCopyDone, pgwirebase.ClientMsgCopyFail:
+				// We're supposed to ignore these messages, per the protocol spec. This
+				// state will happen when an error occurs on the server-side during a copy
+				// operation: the server will send an error and a ready message back to
+				// the client, and must then ignore further copy messages. See:
+				// https://github.com/postgres/postgres/blob/6e1dd2773eb60a6ab87b27b8d9391b756e904ac3/src/backend/tcop/postgres.c#L4295
+				return false, isSimpleQuery, nil
+			default:
+				return false, isSimpleQuery, c.stmtBuf.Push(
+					ctx,
+					sql.SendError{Err: pgwirebase.NewUnrecognizedMsgTypeErr(typ)})
+			}
+		}()
+		if err != nil {
+			log.VEventf(ctx, 1, "pgwire: error processing message: %s", err)
+			if !isSimpleQuery {
+				// In the extended protocol, after seeing an error, we ignore all
+				// messages until receiving a sync.
+				ignoreUntilSync = true
+			}
+			repeatedErrorCount++
+			// If we can't read data because of any one of the following conditions,
+			// then we should break:
+			// 1. the connection was closed.
+			// 2. the context was canceled (e.g. during authentication).
+			// 3. we reached an arbitrary threshold of repeated errors.
+			if netutil.IsClosedConnection(err) ||
+				errors.Is(err, context.Canceled) ||
+				repeatedErrorCount > maxRepeatedErrorCount {
+				break
+			}
+		} else {
+			repeatedErrorCount = 0
+		}
+		if breakLoop {
+			break
+		}
+	}
+
+	// We're done reading data from the client, so make the communication
+	// goroutine stop. Depending on what that goroutine is currently doing (or
+	// blocked on), we cancel and close all the possible channels to make sure we
+	// tickle it in the right way.
+
+	// Signal command processing to stop. It might be the case that the processor
+	// canceled our context and that's how we got here; in that case, this will
+	// be a no-op.
+	c.stmtBuf.Close()
+	// Cancel the processor's context.
+	cancelConn()
+	// In case the authenticator is blocked on waiting for data from the client,
+	// tell it that there's no more data coming. This is a no-op if authentication
+	// was completed already.
+	authenticator.noMorePwdData()
+
+	// Wait for the processor goroutine to finish, if it hasn't already. We're
+	// ignoring the error we get from it, as we have no use for it. It might be a
+	// connection error, or a context cancelation error case this goroutine is the
+	// one that triggered the execution to stop.
+	<-procCh
+
+	if terminateSeen {
+		return
+	}
+	// If we're draining, let the client know by piling on an AdminShutdownError
+	// and flushing the buffer.
+	if s.IsDraining() {
+		// The error here is also sent with pgcode.AdminShutdown, to indicate that
+		// the connection is being closed. Clients are expected to be able to handle
+		// this even when not waiting for a query result. See the discussion at
+		// https://github.com/cockroachdb/cockroach/issues/22630.
+		// NOTE: If a query is canceled due to draining, the conn_executor already
+		// will have sent a QueryCanceled error as a response to the query.
+		log.Ops.Info(ctx, "closing existing connection while server is draining")
+		_ /* err */ = c.writeErr(ctx, newAdminShutdownErr(ErrDrainingExistingConn), &c.writerState.buf)
+		_ /* n */, _ /* err */ = c.writerState.buf.WriteTo(c.conn)
+	}
 }
 
 // readCancelKeyAndCloseConn retrieves the "backend data" key that identifies
